requirements.txt
spotipy
pillow
requests
colorthief
python-dotenv
pywin32